Change bills file to a politics file and start methods for getting politicians
